requirements-host.txt
numpy
pyserial
scipy
sounddevice
soundfile
